+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/IrBindingGraph.kt
@@ -13,6 +13,7 @@ import dev.zacsweers.metro.compiler.tracing.Tracer
 import dev.zacsweers.metro.compiler.tracing.traceNested
 import org.jetbrains.kotlin.cli.common.messages.CompilerMessageSeverity
 import org.jetbrains.kotlin.ir.declarations.IrDeclaration
+import org.jetbrains.kotlin.ir.declarations.IrDeclarationWithName
 import org.jetbrains.kotlin.ir.declarations.IrFunction
 import org.jetbrains.kotlin.ir.declarations.IrProperty
 import org.jetbrains.kotlin.ir.declarations.IrSimpleFunction
@@ -27,6 +28,7 @@ import org.jetbrains.kotlin.ir.types.typeOrNull
 import org.jetbrains.kotlin.ir.types.typeWith
 import org.jetbrains.kotlin.ir.util.dumpKotlinLike
 import org.jetbrains.kotlin.ir.util.fileOrNull
+import org.jetbrains.kotlin.ir.util.fqNameWhenAvailable
 import org.jetbrains.kotlin.ir.util.isSubtypeOf
 import org.jetbrains.kotlin.ir.util.kotlinFqName
 import org.jetbrains.kotlin.ir.util.parentAsClass
@@ -140,7 +142,9 @@ internal class IrBindingGraph(
         }
 
         else -> {
-          error("Unrecognized provider: ${declaration.locationOrNull() ?: ("\n" + declaration.dumpKotlinLike())}")
+          error(
+            "Unrecognized provider: ${declaration.locationOrNull() ?: ("\n" + declaration.dumpKotlinLike())}"
+          )
         }
       }
 
@@ -587,7 +591,7 @@ internal class IrBindingGraph(
   ) {
     if (binding !is IrBinding.ConstructorInjected || !binding.isAssisted) return
 
-    fun reportInvalidBinding(declaration: IrDeclaration?) {
+    fun reportInvalidBinding(declaration: IrDeclarationWithName?) {
       // Look up the assisted factory as a hint
       val assistedFactory =
         bindings.values.find { it is IrBinding.Assisted && it.target.typeKey == binding.typeKey }
@@ -595,7 +599,7 @@ internal class IrBindingGraph(
       val message = buildString {
         append("[Metro/InvalidBinding] ")
         append(
-          "'${binding.typeKey}' uses assisted injection and cannot be injected directly. You must inject a corresponding @AssistedFactory type instead."
+          "'${binding.typeKey}' uses assisted injection and cannot be injected directly into '${declaration?.fqNameWhenAvailable}'. You must inject a corresponding @AssistedFactory type instead."
         )
         if (assistedFactory != null) {
           appendLine()
@@ -606,9 +610,18 @@ internal class IrBindingGraph(
           )
         }
       }
-      metroContext.diagnosticReporter
-        .at(declaration ?: node.sourceGraph)
-        .report(MetroIrErrors.METRO_ERROR, message)
+      val toReport = declaration ?: node.sourceGraph
+      if (toReport.fileOrNull == null) {
+        metroContext.messageCollector.report(
+          CompilerMessageSeverity.ERROR,
+          message,
+          toReport.locationOrNull(),
+        )
+      } else {
+        metroContext.diagnosticReporter
+          .at(declaration ?: node.sourceGraph)
+          .report(MetroIrErrors.METRO_ERROR, message)
+      }
     }
 
     reverseAdjacency[binding.typeKey]?.let { dependents ->
@@ -617,7 +630,7 @@ internal class IrBindingGraph(
         if (dependentBinding !is IrBinding.Assisted) {
           reportInvalidBinding(
             dependentBinding.parametersByKey[binding.typeKey]?.ir?.takeIf {
-              val location = it.location()
+              val location = it.locationOrNull() ?: return@takeIf false
               location.line != 0 || location.column != 0
             } ?: dependentBinding.reportableDeclaration
           )